feat: options logic to until hook
